Tiny struct change

+++ include/eqnet.h
@@ -32,17 +32,17 @@ typedef struct EQNet EQNet;
 ** Initialization and deinitialization
 */
 
-EQNET_API int		EQNet_Init();
-EQNET_API void		EQNet_Close();
-EQNET_API EQNet*	EQNet_Create();
-EQNET_API void		EQNet_Destroy(EQNet*);
+EQNET_API int       EQNet_Init();
+EQNET_API void      EQNet_Close();
+EQNET_API EQNet*    EQNet_Create();
+EQNET_API void      EQNet_Destroy(EQNet*);
 
 enum EQNet_ClientVersion
 {
 	EQNET_CLIENT_TITANIUM
 };
 
-EQNET_API void		EQNet_SetClientVersion(EQNet*, EQNet_ClientVersion version);
+EQNET_API void EQNet_SetClientVersion(EQNet*, EQNet_ClientVersion version);
 
 /*
 ** Fatal Errors
@@ -50,7 +50,7 @@ EQNET_API void		EQNet_SetClientVersion(EQNet*, EQNet_ClientVersion version);
 
 #define EQNET_MAX_ERR_LEN 1024
 
-EQNET_API const char*	EQNet_GetErrorMessage(EQNet*);
+EQNET_API const char* EQNet_GetErrorMessage(EQNet*);
 
 /*
 ** Login
@@ -67,13 +67,13 @@ typedef struct EQNet_Server {
 	uint32_t playerCount;
 } EQNet_Server;
 
-EQNET_API void					EQNet_SetLoginServer(EQNet*, const char* ip, uint16_t port);
-EQNET_API int					EQNet_LoginToServerSelect(EQNet*, const char* username, const char* password);
-EQNET_API int					EQNet_LoginToWorld(EQNet*, EQNet_Server* server);
+EQNET_API void                  EQNet_SetLoginServer(EQNet*, const char* ip, uint16_t port);
+EQNET_API int                   EQNet_LoginToServerSelect(EQNet*, const char* username, const char* password);
+EQNET_API int                   EQNet_LoginToWorld(EQNet*, EQNet_Server* server);
 
-EQNET_API const EQNet_Server*	EQNet_GetServerList(EQNet*, int* count);
-EQNET_API int					EQNet_ServerIsUp(EQNet*, EQNet_Server* server);
-EQNET_API int					EQNet_ServerIsLocked(EQNet*, EQNet_Server* server);
+EQNET_API const EQNet_Server*   EQNet_GetServerList(EQNet*, int* count);
+EQNET_API int                   EQNet_ServerIsUp(EQNet*, EQNet_Server* server);
+EQNET_API int                   EQNet_ServerIsLocked(EQNet*, EQNet_Server* server);
 
 /*
 ** World
@@ -84,21 +84,21 @@ typedef struct EQNet_Guild {
 	char name[64];
 } EQNet_Guild;
 
-EQNET_API const EQNet_Guild*	EQNet_GetGuildList(EQNet*, int* count);
+EQNET_API const EQNet_Guild* EQNet_GetGuildList(EQNet*, int* count);
 
 typedef struct EQNet_Character {
 	char name[64];
-	int level;
-	int charClass;
-	int race;
-	int gender;
-	int deity;
-	int zone;
+	uint8_t level;
+	uint8_t charClass;
+	uint32_t race;
+	uint8_t gender;
+	uint32_t deity;
+	uint32_t zone;
 } EQNet_Character;
 
-EQNET_API const EQNet_Character*	EQNet_GetCharacterList(EQNet*, int* count);
+EQNET_API const EQNet_Character* EQNet_GetCharacterList(EQNet*, int* count);
 
-EQNET_API const char*	EQNet_GetServerShortName(EQNet*);
+EQNET_API const char* EQNet_GetServerShortName(EQNet*);
 
 /*
 ** Events
@@ -108,7 +108,7 @@ typedef struct EQNet_Event {
 	int type;
 } EQNet_Event;
 
-EQNET_API EQNet_Event*	EQNet_Poll(EQNet*);
+EQNET_API EQNet_Event* EQNet_Poll(EQNet*);
 
 enum EQNet_EventType
 {
